requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pydantic==1.10.2
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.4.0
yarl==1.8.1
